feat: update frontend collections, group pages and related components

+++ flutter_app/lib/features/detail/comic_detail_screen.dart
@@ -58,7 +58,7 @@ class _ComicDetailScreenState extends ConsumerState<ComicDetailScreen> {
       final api = ref.read(comicApiProvider);
       await api.updateRating(widget.comicId, rating);
       setState(() {
-        _comic = _comic!.copyWith(rating: rating);
+        _comic = _comic!.copyWith(rating: rating.toDouble());
       });
     } catch (_) {}
   }
